backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
easyocr==1.7.2
opencv-python-headless==4.10.0.84
Pillow==11.0.0
numpy==1.26.4
google-genai>=1.0.0
python-dotenv==1.0.1
httpx>=0.27.0
